Попытался передавать данные в показ анкеты через передачу объекта(не получилось кое где остались куски)

+++ src/main/java/com/pancreatitis/models/Questionnaire.java
@@ -6,8 +6,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.*;
 
 public class Questionnaire {
-    private long id = -1;
-    private long idPatient = -1;
+    private int id = -1;
+    private int idPatient = -1;
     private int idDoctor = -1;
     private int idExpert = -1;
     private String diagnosis;
@@ -56,14 +56,14 @@ public class Questionnaire {
     }
 
     // Геттеры и сеттеры
-    public long getId() { return id; }
-    public void setId(long id) { this.id = id; }
+    public int getId() { return id; }
+    public void setId(int id) { this.id = id; }
 
-    public long getIdPatient() {
+    public int getIdPatient() {
         return idPatient;
     }
 
-    public void setIdPatient(long idPatient) {
+    public void setIdPatient(int idPatient) {
         this.idPatient = idPatient;
     }
